--- source_archive/20221024_AIP_links.csv
@@ -1,17 +1,0 @@
-aip.asp?pg=20&vdate=08SEP2022&ver=1,AIP Book,08SEP2022
-aip.asp?pg=20&vdate=01DEC2022&ver=2,AIP Book,01DEC2022
-aip.asp?pg=60&vdate=16JUN2022&ver=1,AIP Charts,16JUN2022
-aip.asp?pg=60&vdate=01DEC2022&ver=2,AIP Charts,01DEC2022
-aip.asp?pg=50,AIP Supplements (SUP) and Aeronautical  Information Circulars (AIC),
-current/SUP_AIP_Summary/SUPSummary_1022_06OCT2022.pdf,Summary of SUP/AIC,06OCT2022
-current/SUP_AIP_Summary/SUPSummary_1122_03NOV2022.pdf,Summary of SUP/AIC,03NOV2022
-aip.asp?pg=30&vdate=08SEP2022&ver=1,Departure and Approach Procedures (DAP),08SEP2022
-aip.asp?pg=30&vdate=01DEC2022&ver=2,Departure and Approach Procedures (DAP),01DEC2022
-current/dah/DAH_16JUN2022.pdf,Designated Airspace Handbook (DAH),16JUN2022
-pending/dah/DAH_01DEC2022.pdf,Designated Airspace Handbook (DAH),01DEC2022
-aip.asp?pg=40&vdate=08SEP2022&ver=1,En Route Supplement Australia (ERSA),08SEP2022
-aip.asp?pg=40&vdate=01DEC2022&ver=2,En Route Supplement Australia (ERSA),01DEC2022
-current/icao/icao_standards.asp,Differences from ICAO Standards - Recommended Practices and Procedures,02DEC2021
-current/chart/TypeAandBCharts.pdf,Precision Approach Terrain Charts and Type A & Type B Obstacle Charts,
-current/wac/World_Aeronautical_Charts.pdf,World Aeronautical Charts (WAC),
-aip.asp?pg=70,IAIP and chart user guides and supplementary information,
